feat(objective): adds switch component logic
