+++ src/components/Objective/AccordionItem/Menu/wrapper.tsx
@@ -8,7 +8,7 @@ import { objectiveAtomFamily } from '../../../../state/recoil/objective'
 import { teamAtomFamily } from '../../../../state/recoil/team'
 import TreeDotsIcon from '../../../Icon/TreeDots'
 import { GraphQLEffect } from '../../../types'
-import messages from '../Item/messages'
+import messages from '../Item/Button/messages'
 
 import { stopAccordionOpen } from './handlers'
 import { CreateKeyResultOption } from './option-create-key-result'
